perf(dflash): alias predraft Stage 2 outputs onto draft graph buffers
Rename the predraft buffer fields (input_ids, positions, out_cache_loc,
input_embeds) to match the draft decode graph's static slots so
share_buffers() aliases them onto one allocation, giving a zero-copy
hand-off of the embedded block, positions, and cache locs into the draft
replay. Gate the now-aliased draft slots' copy_from_fb off so fill_from
does not self-copy overlapping storage.

Consolidate the three per-step Stage 2 input copies (prefix_lens,
req_pool_indices, verified_id) into one replay_prepare grouped foreach
copy, and migrate the bonus-token verified_id path to int64 so all three
pairs share a dtype group and collapse to a single kernel.

+++ python/sglang/srt/speculative/dflash_info_v2.py
@@ -126,7 +126,7 @@ class DFlashDraftInputV2(SpecInput):
         return cls(
             topk_p=torch.empty((0, 0), device=device, dtype=torch.float32),
             topk_index=torch.empty((0, 0), device=device, dtype=torch.int64),
-            verified_id=torch.empty((0,), device=device, dtype=torch.int32),
+            verified_id=torch.empty((0,), device=device, dtype=torch.int64),
             new_seq_lens=torch.empty((0,), device=device, dtype=torch.int64),
             hidden_states=torch.empty((0, 0), device=device, dtype=torch.float16),
             verify_done=None,
